Remove BC4 Flavor (#667)
