+++ api-tester/build.gradle
@@ -14,5 +14,4 @@ dependencies {
     implementation project(path: ':public')
     implementation project(path: ':purchases')
     implementation "androidx.annotation:annotation:$annotationVersion"
-    unityIAPCompileOnly "com.android.billingclient:billing:$billing4Version"
 }
