+++ base-application.gradle
@@ -21,15 +21,6 @@ android {
     kotlinOptions {
         jvmTarget = JavaVersion.VERSION_1_8.toString()
     }
-    flavorDimensions "dependencyVersions"
-    productFlavors {
-        unityIAP {
-            dimension "dependencyVersions"
-        }
-        latestDependencies {
-            dimension "dependencyVersions"
-        }
-    }
 }
 
 dependencies {
