+++ feature/google/build.gradle
@@ -14,8 +14,7 @@ dependencies {
     implementation 'com.google.android.gms:play-services-ads-identifier:17.0.1'
     implementation "androidx.core:core-ktx:$androidxCoreVersion"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-    latestDependenciesImplementation "com.android.billingclient:billing:$billing5Version"
-    unityIAPCompileOnly "com.android.billingclient:billing:$billing4Version"
+    implementation "com.android.billingclient:billing:$billingVersion"
 
     testImplementation project(":test-utils")
     testImplementation "androidx.test:core:$testLibrariesVersion"
@@ -26,6 +25,5 @@ dependencies {
     testImplementation "io.mockk:mockk:$mockkVersion"
     testImplementation "org.assertj:assertj-core:$assertJVersion"
     testImplementation 'com.google.android.gms:play-services-ads-identifier:17.0.1'
-    testUnityIAPImplementation "com.android.billingclient:billing:$billing4Version"
 
 }
