+++ gradle.properties
@@ -25,6 +25,6 @@ purchaseTesterSupportedStores=amazon,google
 
 #Do not sign releases. When calling uploadArchives pass -PRELEASE_SIGNING_ENABLED=true
 RELEASE_SIGNING_ENABLED=false
-ANDROID_VARIANT_TO_PUBLISH=latestDependenciesRelease
+ANDROID_VARIANT_TO_PUBLISH=release
 SONATYPE_HOST=DEFAULT
 SONATYPE_AUTOMATIC_RELEASE=true
