+++ integration-tests/build.gradle
@@ -46,7 +46,6 @@ dependencies {
     androidTestImplementation 'androidx.test:rules:1.4.0'
     androidTestImplementation 'androidx.test.ext:junit-ktx:1.1.3'
     androidTestImplementation 'org.assertj:assertj-core:3.22.0'
-    androidTestUnityIAPImplementation "com.android.billingclient:billing:$billing4Version"
 }
 
 def obtainTestBuildType() {
