+++ public/build.gradle
@@ -11,8 +11,7 @@ dependencies {
     implementation project(":strings")
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
     implementation "androidx.core:core-ktx:$androidxCoreVersion"
-    latestDependenciesImplementation "com.android.billingclient:billing:$billing5Version"
-    unityIAPCompileOnly "com.android.billingclient:billing:$billing4Version"
+    implementation "com.android.billingclient:billing:$billingVersion"
     testImplementation project(":test-utils")
     testImplementation "androidx.test:core:$testLibrariesVersion"
     testImplementation "androidx.test:runner:$testLibrariesVersion"
