+++ purchases/build.gradle
@@ -18,8 +18,7 @@ dependencies {
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
     implementation "androidx.annotation:annotation:$annotationVersion"
-    latestDependenciesApi "com.android.billingclient:billing:$billing5Version"
-    unityIAPCompileOnly "com.android.billingclient:billing:$billing4Version"
+    api "com.android.billingclient:billing:$billingVersion"
     implementation "androidx.lifecycle:lifecycle-runtime:$lifecycleVersion"
     implementation "androidx.lifecycle:lifecycle-common:$lifecycleVersion"
     implementation "androidx.lifecycle:lifecycle-process:$lifecycleVersion"
@@ -32,8 +31,7 @@ dependencies {
     testImplementation "org.robolectric:robolectric:$robolectricVersion"
     testImplementation "io.mockk:mockk:$mockkVersion"
     testImplementation "org.assertj:assertj-core:$assertJVersion"
-    testLatestDependenciesImplementation "com.android.billingclient:billing:$billing5Version"
-    testUnityIAPImplementation "com.android.billingclient:billing:$billing4Version"
+    testImplementation "com.android.billingclient:billing:$billingVersion"
 }
 
 tasks.dokkaHtmlPartial.configure {
