+++ test-utils/build.gradle
@@ -10,8 +10,7 @@ dependencies {
     implementation project(":feature:google")
     implementation project(":utils")
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-    latestDependenciesImplementation "com.android.billingclient:billing:$billing5Version"
-    unityIAPCompileOnly "com.android.billingclient:billing:$billing4Version"
+    implementation "com.android.billingclient:billing:$billingVersion"
     implementation "org.assertj:assertj-core:$assertJVersion"
     implementation "io.mockk:mockk:$mockkVersion"
     implementation "androidx.test.ext:junit:$testJUnitVersion"
